+++ src/tests/e2e/remote_runner.rs
@@ -235,6 +235,7 @@ fn read_from_channel(
                 })));
                 let sixel_image_store = Rc::new(RefCell::new(SixelImageStore::default()));
                 let debug = false;
+                let arrow_fonts = true;
                 let mut terminal_output = TerminalPane::new(
                     0,
                     pane_geom,
@@ -249,6 +250,7 @@ fn read_from_channel(
                     None,
                     None,
                     debug,
+                    arrow_fonts,
                 ); // 0 is the pane index
                 loop {
                     if !should_keep_running.load(Ordering::SeqCst) {