+++ zellij-server/src/panes/unit/terminal_pane_tests.rs
@@ -37,6 +37,7 @@ pub fn scrolling_inside_a_pane() {
     let terminal_emulator_colors = Rc::new(RefCell::new(Palette::default()));
     let terminal_emulator_color_codes = Rc::new(RefCell::new(HashMap::new()));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -51,6 +52,7 @@ pub fn scrolling_inside_a_pane() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let mut text_to_fill_pane = String::new();
     for i in 0..30 {
@@ -81,6 +83,7 @@ pub fn sixel_image_inside_terminal_pane() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -95,6 +98,7 @@ pub fn sixel_image_inside_terminal_pane() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let sixel_image_bytes = "\u{1b}Pq
         #0;2;0;0;0#1;2;100;100;0#2;2;0;100;0
@@ -125,6 +129,7 @@ pub fn partial_sixel_image_inside_terminal_pane() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -139,6 +144,7 @@ pub fn partial_sixel_image_inside_terminal_pane() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let pane_content = read_fixture("sixel-image-500px.six");
     terminal_pane.handle_pty_bytes(pane_content);
@@ -163,6 +169,7 @@ pub fn overflowing_sixel_image_inside_terminal_pane() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -177,6 +184,7 @@ pub fn overflowing_sixel_image_inside_terminal_pane() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let pane_content = read_fixture("sixel-image-500px.six");
     terminal_pane.handle_pty_bytes(pane_content);
@@ -200,6 +208,7 @@ pub fn scrolling_through_a_sixel_image() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -214,6 +223,7 @@ pub fn scrolling_through_a_sixel_image() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let mut text_to_fill_pane = String::new();
     for i in 0..30 {
@@ -248,6 +258,7 @@ pub fn multiple_sixel_images_in_pane() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -262,6 +273,7 @@ pub fn multiple_sixel_images_in_pane() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let mut text_to_fill_pane = String::new();
     for i in 0..5 {
@@ -294,6 +306,7 @@ pub fn resizing_pane_with_sixel_images() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -308,6 +321,7 @@ pub fn resizing_pane_with_sixel_images() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let mut text_to_fill_pane = String::new();
     for i in 0..5 {
@@ -343,6 +357,7 @@ pub fn changing_character_cell_size_with_sixel_images() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -357,6 +372,7 @@ pub fn changing_character_cell_size_with_sixel_images() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
     let mut text_to_fill_pane = String::new();
     for i in 0..5 {
@@ -397,6 +413,7 @@ pub fn keep_working_after_corrupted_sixel_image() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -411,6 +428,7 @@ pub fn keep_working_after_corrupted_sixel_image() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
 
     let sixel_image_bytes = "\u{1b}PI AM CORRUPTED BWAHAHAq
@@ -449,6 +467,7 @@ pub fn pane_with_frame_position_is_on_frame() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -463,6 +482,7 @@ pub fn pane_with_frame_position_is_on_frame() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
 
     terminal_pane.set_content_offset(Offset::frame(1));
@@ -537,6 +557,7 @@ pub fn pane_with_bottom_and_right_borders_position_is_on_frame() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -551,6 +572,7 @@ pub fn pane_with_bottom_and_right_borders_position_is_on_frame() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
 
     terminal_pane.set_content_offset(Offset::shift(1, 1));
@@ -625,6 +647,7 @@ pub fn frameless_pane_position_is_on_frame() {
         height: 21,
     })));
     let debug = false;
+    let arrow_fonts = true;
     let mut terminal_pane = TerminalPane::new(
         pid,
         fake_win_size,
@@ -639,6 +662,7 @@ pub fn frameless_pane_position_is_on_frame() {
         None,
         None,
         debug,
+        arrow_fonts,
     ); // 0 is the pane index
 
     terminal_pane.set_content_offset(Offset::default());