+++ zellij-server/src/tab/layout_applier.rs
@@ -41,6 +41,7 @@ pub struct LayoutApplier<'a> {
     focus_pane_id: &'a mut Option<PaneId>,
     os_api: Box<dyn ServerOsApi>,
     debug: bool,
+    arrow_fonts: bool,
 }
 
 impl<'a> LayoutApplier<'a> {
@@ -61,6 +62,7 @@ impl<'a> LayoutApplier<'a> {
         focus_pane_id: &'a mut Option<PaneId>,
         os_api: &Box<dyn ServerOsApi>,
         debug: bool,
+        arrow_fonts: bool,
     ) -> Self {
         let viewport = viewport.clone();
         let senders = senders.clone();
@@ -90,6 +92,7 @@ impl<'a> LayoutApplier<'a> {
             focus_pane_id,
             os_api,
             debug,
+            arrow_fonts,
         }
     }
     pub fn apply_layout(
@@ -256,6 +259,7 @@ impl<'a> LayoutApplier<'a> {
                             self.style,
                             layout.run.clone(),
                             self.debug,
+                            self.arrow_fonts,
                         );
                         if let Some(pane_initial_contents) = &layout.pane_initial_contents {
                             new_plugin.handle_pty_bytes(pane_initial_contents.as_bytes().into());
@@ -292,6 +296,7 @@ impl<'a> LayoutApplier<'a> {
                                 initial_title,
                                 layout.run.clone(),
                                 self.debug,
+                                self.arrow_fonts,
                             );
                             if let Some(pane_initial_contents) = &layout.pane_initial_contents {
                                 new_pane.handle_pty_bytes(pane_initial_contents.as_bytes().into());
@@ -381,6 +386,7 @@ impl<'a> LayoutApplier<'a> {
                     self.style,
                     floating_pane_layout.run.clone(),
                     self.debug,
+                    self.arrow_fonts,
                 );
                 if let Some(pane_initial_contents) = &floating_pane_layout.pane_initial_contents {
                     new_pane.handle_pty_bytes(pane_initial_contents.as_bytes().into());
@@ -420,6 +426,7 @@ impl<'a> LayoutApplier<'a> {
                     initial_title,
                     floating_pane_layout.run.clone(),
                     self.debug,
+                    self.arrow_fonts,
                 );
                 if let Some(pane_initial_contents) = &floating_pane_layout.pane_initial_contents {
                     new_pane.handle_pty_bytes(pane_initial_contents.as_bytes().into());