+++ zellij-tile/src/ui_components/mod.rs
@@ -1,0 +1,12 @@
+mod nested_list;
+mod ribbon;
+mod table;
+mod text;
+
+pub use zellij_utils::plugin_api;
+pub use zellij_utils::prost::{self, *};
+
+pub use nested_list::*;
+pub use ribbon::*;
+pub use table::*;
+pub use text::*;