+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * feat(ui): cyberpunk themes (https://github.com/zellij-org/zellij/pull/2868)
 * feat(ux): add ESC option to drop to shell in command panes (https://github.com/zellij-org/zellij/pull/2872)
 * feat(plugins): allow plugins to make web requests behind a permission (https://github.com/zellij-org/zellij/pull/2879)
+* feat(plugins): UI components for plugins (https://github.com/zellij-org/zellij/pull/2898)
+* feat(plugins): load plugins from the web (https://github.com/zellij-org/zellij/pull/2863)
 
 ## [0.38.2] - 2023-09-15
 * fix(terminal): wrap lines in alternate screen mode when adding characters (https://github.com/zellij-org/zellij/pull/2789)